SCC API check: server
