+++ pkg/security/scc/byrestrictions_test.go
@@ -1,4 +1,4 @@
-package admission
+package scc
 
 import (
 	"testing"
